menu actualizado con correcciones
