+++ uade_battleship/instructions.py
@@ -1,6 +1,6 @@
 import os
 import pygame, sys
-from uade_battleship.main_menu import main_menu
+from .ui import Button
 
 # Inicializar Pygame
 pygame.init()
@@ -22,52 +22,10 @@ def get_font(size):
 
 
 # Clase para los botones
-class Button:
-    def __init__(
-        self,
-        image,
-        pos,
-        text_input,
-        font,
-        base_color,
-        hovering_color,
-        bg_color,
-        border_color,
-    ):
-        self.image = image
-        self.x_pos = pos[0]
-        self.y_pos = pos[1]
-        self.font = font
-        self.base_color = base_color
-        self.hovering_color = hovering_color
-        self.bg_color = bg_color
-        self.border_color = border_color
-        self.text_input = text_input
-        self.text = self.font.render(self.text_input, True, self.base_color)
-
-        if self.image is None:
-            self.image = self.text
-        self.rect = self.image.get_rect(center=(self.x_pos, self.y_pos))
-        self.text_rect = self.text.get_rect(center=(self.x_pos, self.y_pos))
-
-    def update(self, screen):
-        pygame.draw.rect(screen, self.bg_color, self.rect)
-        pygame.draw.rect(screen, self.border_color, self.rect, 2)  # Borde del botón
-        screen.blit(self.image, self.rect)
-        screen.blit(self.text, self.text_rect)
-
-    def check_for_input(self, position):
-        return self.rect.collidepoint(position)
-
-    def change_color(self, position):
-        if self.rect.collidepoint(position):
-            self.text = self.font.render(self.text_input, True, self.hovering_color)
-        else:
-            self.text = self.font.render(self.text_input, True, self.base_color)
 
 
 # Función para mostrar las reglas del juego
-def reglas():
+def instructions():
     while True:
         screen.fill(DARK_BLUE)
 
@@ -97,7 +55,7 @@ def reglas():
 
             if event.type == pygame.MOUSEBUTTONDOWN:
                 if back_button.check_for_input(pygame.mouse.get_pos()):
-                    main_menu()
+                    return
 
         back_button.change_color(pygame.mouse.get_pos())
