+++ uade_battleship/ui/Button.py
@@ -1,0 +1,45 @@
+import pygame
+
+
+class Button:
+    def __init__(
+        self,
+        image,
+        pos,
+        text_input,
+        font,
+        base_color,
+        hovering_color,
+        bg_color,
+        border_color,
+    ):
+        self.image = image
+        self.x_pos = pos[0]
+        self.y_pos = pos[1]
+        self.font = font
+        self.base_color = base_color
+        self.hovering_color = hovering_color
+        self.bg_color = bg_color
+        self.border_color = border_color
+        self.text_input = text_input
+        self.text = self.font.render(self.text_input, True, self.base_color)
+
+        if self.image is None:
+            self.image = self.text
+        self.rect = self.image.get_rect(center=(self.x_pos, self.y_pos))
+        self.text_rect = self.text.get_rect(center=(self.x_pos, self.y_pos))
+
+    def update(self, screen):
+        pygame.draw.rect(screen, self.bg_color, self.rect)
+        pygame.draw.rect(screen, self.border_color, self.rect, 2)  # Borde del botón
+        screen.blit(self.image, self.rect)
+        screen.blit(self.text, self.text_rect)
+
+    def check_for_input(self, position):
+        return self.rect.collidepoint(position)
+
+    def change_color(self, position):
+        if self.rect.collidepoint(position):
+            self.text = self.font.render(self.text_input, True, self.hovering_color)
+        else:
+            self.text = self.font.render(self.text_input, True, self.base_color)
