+++ uade_battleship/ui/__init__.py
@@ -1,0 +1,1 @@
+from .Button import Button
